move GroupByExecutor from dataset to excutor
