+++ server/src/main/java/org/apache/iotdb/db/query/executor/groupby/LocalGroupByExecutor.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.query.dataset.groupby;
+package org.apache.iotdb.db.query.executor.groupby;
 
 import org.apache.iotdb.db.engine.querycontext.QueryDataSource;
 import org.apache.iotdb.db.exception.StorageEngineException;
